INTDEV-733: Split scanning and rank planning out of card-tree.ts
card-tree.ts held three jobs at once: the tree itself, the recursive disk
sweep that fills it, and the lexorank arithmetic that decides what a
reorder costs. The two supporting jobs neither touch the tree's state nor
need to, so they move to their own files and the tree calls them.

card-tree-scan.ts owns the on-disk shape of a card: the folder and file
names, StoredCard/StoredAttachment, the sweep (scanCardTree) and
normalizedMetadata, which is what turns metadata read off disk into the
frozen stored form. normalizedMetadata travels with StoredCard because it
is that type's constructor; the tree's writers import it back.

rank-plan.ts owns the planners - planAfter, planFirst, planRebalance,
planRebalanceSubtree and withUsableRanks - as pure functions over
{ key, rank? }[]. The tree keeps everything that reads or writes its own
state: siblingsUnder, rankOf, rankBlock, applyRanks, writeRank. Two
consequences of making the planners pure: siblingsUnder returns the
sibling ranks alongside the keys rather than the planners reaching back
for them, and planAfter's assertion that the moved card exists moves up
into reorderAfter, which is where the throw was documented already.

No behavior change; the public surface is unchanged and no test changed.
Two deliberate differences, neither observable: planRebalanceSubtree
recurses into a childless card instead of testing for children first
(rebalanceRanks(0) is []), and the metadata-parse error log gains a 'tree'
field, which is the scan's only use for the tree name it is handed.
